LM-47: Do not remove all invitations when joining a group

+++ backend/LetsMeatAPI/Controllers/GroupsController.cs
@@ -139,11 +139,9 @@ namespace LetsMeatAPI.Controllers {
         return NotFound();
       var user = await _context.Users.FindAsync(userId);
       grp.Users.Add(user);
-      foreach(var inv in from inv in _context.Invitations
-                         where inv.ToId == userId
-                         select inv) {
+      var inv = await _context.Invitations.FindAsync(userId, grp.Id);
+      if(inv != null)
         _context.Invitations.Remove(inv);
-      }
       try {
         await _context.SaveChangesAsync();
       } catch(DbUpdateException ex) {
